requirements.txt
selenium
bs4
numpy
pandas
requests
boto3
azure-storage-blob==0.37.0
pyvirtualdisplay
xvfbwrapper
awscli